[morph]: move RepeatMethod enum back to FE as it is not used in library
